+++ src/js/plyr.js
@@ -263,7 +263,7 @@ class Plyr {
 
         // Wrap media
         if (!is.element(this.elements.container)) {
-            this.elements.container = createElement('div');
+            this.elements.container = createElement('div', { tabindex: 0 });
             wrap(this.media, this.elements.container);
         }